move ButtonProps to types.ts
